analytics_service/requirements.txt
mysql-connector-python
statistics
pymongo